comp-345-ski-resort/requirements.txt
mysql-connector-python>==9.5.0
tabulate>=0.9.0